refactor: move main cmd and configs setup
