+++ src/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+func LoadConfig() {
+	LoadEnv()
+	ConnectDb()
+}
